move BuilderBase to libutils and rename to PrivateImplementation
